PcJsonクラスの追加とそれに伴うPcクラスの修正
